Updated new approach.
Extract information of a whole directory (program), not an individual file.
Based on the list of interesting files (the files that are not managed
by APT), extract their directories (program's folders).

In each program's folder, find a 'reprentative' file, this file will be
used to:
  - Identify modified date-time of the program in local machine
  - Find the online github (if available) to get info about latest releases.

Currently, this 'representative' file is chosen by its file path.
The file with longest filepath will be chose.