Added file for getting information from Internet.
This script will get a number of files that have longest file paths
inside the provided program dir, from the interesting list.

Then, use GitHub API to search for those filepaths to find their
public repos on internet

After that, it filters the results: get only the repo that appear in
all returned results. (Since each search with individual file path
returns more than one possible repo. The one we are looking for is
the one that exist in all the returned results.)